+++ netcheck/checks/tcp.py
@@ -1,6 +1,7 @@
 import datetime
 import logging
 import socket
+from typing import Optional
 
 logger = logging.getLogger("netcheck.tcp")
 DEFAULT_TCP_VALIDATION_RULE = """
@@ -8,13 +9,17 @@ data.connected == true
 """
 
 
-def tcp_check(host: str, port: int, timeout: float = 5) -> dict:
+def tcp_check(
+    host: str, port: int, timeout: float = 5, source_ip: Optional[str] = None
+) -> dict:
     test_spec = {
         "type": "tcp",
         "host": host,
         "port": port,
         "timeout": timeout,
     }
+    if source_ip is not None:
+        test_spec["source-ip"] = source_ip
 
     result_data = {
         "startTimestamp": datetime.datetime.utcnow().isoformat(),
@@ -22,8 +27,12 @@ def tcp_check(host: str, port: int, timeout: float = 5) -> dict:
 
     output = {"spec": test_spec, "data": result_data}
 
+    source_address = (source_ip, 0) if source_ip is not None else None
+
     try:
-        with socket.create_connection((host, port), timeout=timeout):
+        with socket.create_connection(
+            (host, port), timeout=timeout, source_address=source_address
+        ):
             result_data["connected"] = True
             result_data["error"] = None
     except socket.timeout: